+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * #17 Upgrade bootique-kafka-client to Kafka 2.0
 * #18 bootique-kafka-streams: a new module to support Kafka streams
 * #19 Rename the project to "bootique-kafka"
+* #21 "kafkaclient.producer.lingerMs" - rename to "linger" and use Duration type
 
 ## 0.25